Move functions to lib
